fix: replaced mcp modal with mcp menu like repo menu in prompt section
