+++ packages/web/app/api/cron/agent-lifecycle/route.ts
@@ -28,9 +28,10 @@ export const maxDuration = 300
 // Timeouts
 // =============================================================================
 
-const INTERACTIVE_INACTIVITY_TIMEOUT = 10 // minutes
-const INTERACTIVE_HARD_TIMEOUT = 25 // minutes
-const SCHEDULED_HARD_TIMEOUT = 20 // minutes
+const INTERACTIVE_INACTIVITY_TIMEOUT =  7 // minutes
+const INTERACTIVE_HARD_TIMEOUT = 10 // minutes
+const SCHEDULED_HARD_TIMEOUT = 12
+ // minutes
 
 // =============================================================================
 // Types
